Rename `rustdoc::html::render::cache` to `search_index`
The old name wasn't very clear, while the new one makes it clear that
this is the code responsible for creating the search index.
