+++ src/librustdoc/html/render/search_index.rs
